# openapi/audio-studio-v1.json
{
  "components": {
    "schemas": {
...
      "BatchJobCreate": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Columns",
            "type": "object"
          },
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "format": {
            "default": "mp3",
            "title": "Format",
            "type": "string"
          },
          "instruction": {
            "default": "",
            "title": "Instruction",
            "type": "string"
          },
          "language": {
            "default": "",
            "title": "Language",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "pitch": {
            "default": 1,
            "title": "Pitch",
            "type": "number"
          },
          "rate": {
            "default": 1,
            "title": "Rate",
            "type": "number"
          },
          "token": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          },
          "voice": {
            "minLength": 1,
            "title": "Voice",
            "type": "string"
          },
          "voice_identity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Identity Id"
          },
          "volume": {
            "default": 50,
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "token",
          "columns",
          "voice",
          "engine",
          "model"
        ],
        "title": "BatchJobCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RenderJobCreate": {
        "additionalProperties": false,
        "properties": {
          "operation": {
            "enum": [
              "preview",
              "export"
            ],
            "title": "Operation",
            "type": "string"
          },
          "production_id": {
            "exclusiveMinimum": 0.0,
            "title": "Production Id",
            "type": "integer"
          }
        },
        "required": [
          "production_id",
          "operation"
        ],
        "title": "RenderJobCreate",
        "type": "object"
      },
...
      "SpeechJobCreate": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "format": {
            "default": "mp3",
            "title": "Format",
            "type": "string"
          },
          "insert_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insert At"
          },
          "instruction": {
            "default": "",
            "title": "Instruction",
            "type": "string"
          },
          "language": {
            "default": "Auto",
            "title": "Language",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "operation": {
            "default": "create",
            "enum": [
              "create",
              "regenerate",
              "render_draft"
            ],
            "title": "Operation",
            "type": "string"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "pitch": {
            "default": 1,
            "title": "Pitch",
            "type": "number"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "rate": {
            "default": 1,
            "title": "Rate",
            "type": "number"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "speech_mode": {
            "default": "exact",
            "title": "Speech Mode",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "text_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Raw"
          },
          "text_shaped": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Shaped"
          },
          "text_state": {
            "default": "raw",
            "title": "Text State",
            "type": "string"
          },
          "text_tagged": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Tagged"
          },
          "voice": {
            "minLength": 1,
            "title": "Voice",
            "type": "string"
          },
          "voice_identity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Identity Id"
          },
          "volume": {
            "default": 50,
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "text",
          "voice",
          "engine",
          "model"
        ],
        "title": "SpeechJobCreate",
        "type": "object"
      },
...
      "TextJobCreate": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "density": {
            "default": "normal",
            "title": "Density",
            "type": "string"
          },
          "engine": {
            "default": "audio",
            "title": "Engine",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "operation": {
            "enum": [
              "shape",
              "tag"
            ],
            "title": "Operation",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "operation",
          "text"
        ],
        "title": "TextJobCreate",
        "type": "object"
      },
      "TranscriptionJobCreate": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "duration_ms": {
            "default": 0,
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "enable_itn": {
            "default": false,
            "title": "Enable Itn",
            "type": "boolean"
          },
          "file": {
            "default": "",
            "title": "File",
            "type": "string"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          },
          "language": {
            "default": "",
            "title": "Language",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "playable": {
            "default": "",
            "title": "Playable",
            "type": "string"
          },
          "size_bytes": {
            "default": 0,
            "maximum": 500000000.0,
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "vocabulary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocabulary Id"
          }
        },
        "title": "TranscriptionJobCreate",
        "type": "object"
      },
      "TranslationJobCreate": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "quality": {
            "default": "fast",
            "title": "Quality",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "target": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target"
        ],
        "title": "TranslationJobCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VORVN Audio Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/batch": {
      "post": {
        "operationId": "createBatchJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createbatchjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/render": {
      "post": {
        "operationId": "createRenderJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createrenderjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Render Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/speech": {
      "post": {
        "operationId": "createSpeechJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createspeechjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Speech Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/text": {
      "post": {
        "operationId": "createTextJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createtextjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Text Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/transcription": {
      "post": {
        "operationId": "createTranscriptionJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createtranscriptionjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Transcription Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/translation": {
      "post": {
        "operationId": "createTranslationJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createtranslationjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Translation Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Canceljob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "listJobEvents",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listjobevents",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}